+++ ArchiSteamFarm/Json/Steam.cs
@@ -38,7 +38,7 @@ namespace ArchiSteamFarm.Json {
 			public const uint SteamAppID = 753;
 
 			[PublicAPI]
-			public const uint SteamCommunityContextID = 6;
+			public const ulong SteamCommunityContextID = 6;
 
 			[PublicAPI]
 			public uint Amount { get; internal set; }
@@ -53,7 +53,7 @@ namespace ArchiSteamFarm.Json {
 			public ulong ClassID { get; private set; }
 
 			[PublicAPI]
-			public uint ContextID { get; private set; }
+			public ulong ContextID { get; private set; }
 
 			[PublicAPI]
 			public bool Marketable { get; internal set; }
@@ -146,7 +146,7 @@ namespace ArchiSteamFarm.Json {
 						return;
 					}
 
-					if (!uint.TryParse(value, out uint contextID) || (contextID == 0)) {
+					if (!ulong.TryParse(value, out ulong contextID) || (contextID == 0)) {
 						ASF.ArchiLogger.LogNullError(nameof(contextID));
 
 						return;
@@ -164,7 +164,7 @@ namespace ArchiSteamFarm.Json {
 			}
 
 			// Constructed from trades being received or plugins
-			public Asset(uint appID, uint contextID, ulong classID, uint amount, bool marketable = true, uint realAppID = 0, EType type = EType.Unknown, ERarity rarity = ERarity.Unknown) {
+			public Asset(uint appID, ulong contextID, ulong classID, uint amount, bool marketable = true, uint realAppID = 0, EType type = EType.Unknown, ERarity rarity = ERarity.Unknown) {
 				if ((appID == 0) || (contextID == 0) || (classID == 0) || (amount == 0)) {
 					throw new ArgumentNullException(nameof(appID) + " || " + nameof(contextID) + " || " + nameof(classID) + " || " + nameof(amount));
 				}
